Avance 2, correcciones y mapping

+++ evaluation-service/src/main/java/com/davidsonperez/evalservice/evaluationservice/data/entity/Student.java
@@ -30,6 +30,6 @@ public class Student implements Serializable {
     private String name;
     @Column(nullable = false)
     private String email;
-    private Float qualification;
+    private Double qualification;
     private Exam exam;
 }
